Change to GPU writing unpadded Partial Accelerations.
This massively simplifies Finalize and generally yields fairly
simple code.  Separated the Partial Accelerations from the rest
of the SIC storage.

But this version has an intermittent problem.  It crashes in Spiral
after ~10 steps, but appears to be correct until then.  The problem
appears to trigger an overly large drift.  Maybe a race condition?