Found the bug, hacked fix.
The problem was that the sink particles aren't unique in a Z modulus Set,
because of the periodic wrap if CPD%WIDTH>0.  While the padded partial
accelerations are held distinctly, there can be 2 for some particles
and 0 for others.  So we can't just memcpy into a PartialAccSlab that
has exactly 1 entry per particle.

The current commit fixes this by initializing the whole Partial
array to zero, and then adding instead of copying.

This is actually both a problem and an opportunity.  As written,
it is needlessly slow.  It highlights that there is not a strong
reason to limit a SIC to a Z modulus -- if the particles aren't going
to be distinct, what's the point?

But if we changed to having a smaller range of Y and all Z in a SIC,
then we could coadd onto the one AccSlab directly from pinned memory.
And probably still get good cache re-use, because we'd be proceeding
down the line of Z's, much as the current Finalize does.

The downside is that a full Y skewer of data will require more
SourcePencil padding.  But GPUs have more memory now, so maybe
we should just do this?