Trying to avoid some sorting in SO
This uses the total number of remaining particles to set a maximum
distance and then cuts the particle list before it gets to the sort,
thereby reducing the amount of sorting.  This appears to speed up
the overall SO work by about 40%.

But there may be more tricks possible.  Starting work on doing this
with histograms.