Histogram-based SO code is working
This version radically reduces the sorting load by
a) clipping all particles that exceed the radius of the input mass,
b) doing a first-pass histogram and confining where the radius could be.

The answers match exactly to the "standard" code.  The time in the SO
routine is greatly reduced, and the number of sorted elements is down by
a factor of 50.

That said, I will next look to tune where the transition should be
between the two methods, because I bet that sorting is faster than
this histogramming for smaller groups.