Fixing bugs in double-precision group finding.
This version runs for both SO and FOF in both single and double-precision.

If --enable-double is set, then the group finding will run in single
precision if --enable-avx-fof is set, and in double precision if not.

This now allocates a new buffer for permuting particles in cell groups,
rather than reusing the p array.  Re-use was causing bugs in mixed precision,
and I think in general had some risks as I wasn't protecting the AVX
ghost particles enough.  Cells are initially allocated to about 60K
particle maxsize, and rarely if ever exceed that, so this is 1 MB per
thread in single precision, twice that in double.  In principle, this
could be avoided, but the memory usage is quite modest.