Fixing numerous compile bugs; ExpandPathName bug.
The compile bugs were simple variable names issues.

However, the ExpandPathName bug is weird.  The call to the
system function realpath() is failing in -O3, unless we include
an explicit use of the errno return value.  This reports as a
buffer overflow.  We suspect a bug in g++ libraries on ted,
as this problem wasn't found in icc.