Converting MPI code in manifest to deal with sizes >2 GB
This changes the request structure into parts of size SIZE_MPI.
This implies that a single arena/list of the manifest may span
multiple parts, so we can't delete until all of the manifest is
received.  This is probably harmless: we weren't guaranteeing
that the Send and Receive manifests would arrive at the same
time anyways.

In some ways, this may have simplified the code a bit.