Adjustments to swizzling
The code for reflecting one octant into all of the rest was not choosing a good
memory access pattern.  The closest spaced data (z or k) was the outermost loop.
I changed it to be the innermost loop, and moreover to pre-compute the indices
involved with k=0 and then do ++ or -- after that.

A similar change for the cyclic translation.  Here the memory pattern
wasn't bad, but there's no need to keep recomputing the index in the
innermost dimension.  So now it just copies the k-skewer with a periodic
wrap halfway through.

This version compiles.