Speed improvements in InCoreConvolve.
This changes the ordering of the real and imaginary derivatives
of the convolution, and then multiplies by I after the imag, so
that we can keep accumulating in the same tcache.  This avoids
touching the main Taylor memory twice.  Saved about 10% on ted.

Then changed to avoid a modulus unwrapping of xyz in the derivative
loading.  This was happening for every cell and every multipole.
Saved nearly 15% on the ICC work.