Fixed periodic wrap dependency issue
This version runs on a single node, provided one isn't microstepping.
It gets the same L0 mass function as the two-sided version.
Also, the extra time to defer groups and then clear those referrals
appears to be quite fast, percent-level changes in the run time.
It seems that the time to construct all of the cell groups and links
is much more than the time to traverse the graphs.

Microstepping would break this: the fact that group finding doesn't
happen at one time in a slab means that the output needs to happen
before any of the slabs get microstepped, and yet the output is supposed
to know about the L0 membership to separate the field particles.