Option: output L0 but non-L1 particles in L1 subsample files
This is commented out, but I expect we'll need to go in this direction.
The issue is that we need to complete all of the L0 work before discarding
the groups, including the Microstepping.  The Output step must be after all
GroupFinding.  Yet the subsamples outputs were based on L1, not L0.