Moving reset_L01_bits() to merge
We had a bug that these bits were getting set as the
CellGroups were being found.  That means that they were
getting written out to the WriteState and would be persistent
into future steps.  As our Output now uses the bits, this
would cause L0 particles to get skipped in the output of
future steps that don't involve group finding.  E.g. light cones.

Now the reset happens as we're forming the MergeAuxSlab, so
these bits will just not get outputted.