Changing the scalings of the min_central mass for L2.
The fixed choice of 35 particles was causing us to derive a small density for the
late-time L2, compared to the expected density at the edge of groups.
This was probably causing a lot of false positive searches at the edge of
groups, where we'd have to compute the distances and sort the first cell, only
to find that we were below the threshold.  So the group would find a small mass
and then go to the next false positive.

This is related to the fact that we don't put a floor on the min_central density
at the enclosed density threshold.