Adding timing to each manifest for MPI lag
This measures the time between the initiation of the major Isends/Irecv
and the recognition that they're all done.  Result is put in the logs.
Note that this time is an overestimate of the actual transfer time, because
we only check for completion at particular points in the timestep loop.