Combining four different timing/summary files into one
The stats from ParallelConvolve are now written into a string buffer, then appended to the .time buffer.

Similarly, the GLOG that was going to .groupstats is now a string buffer that gets appended.

And the .slabtimes outputs, which were in binary, are now in ASCII at the end of that buffer.