Rely on LC bits to groom unwanted duplicates
Our attempts to match the precision of the drift, so that we had perfect
reproducibility between one time step and the next wasn't working at the
part per million level, so we'll just go back to the original bits.

This still has the convention that particles that go through the periodic
wrap reappear with the bits cleared.