Changing tolerance on LC comparison to single precision
We had allowed for a 1e-13 variation, but changing to 1e-6.

The concern is that a particle in this step and the same particle
in the next step may have its position altered at 1e-7 of a cell
width due to single-precision disagreements in the rebinning of the
particle into a new cell.  We are measuring the particle position
to the width of the LC annulus, which is typically 5-10 cells.
That comparison is all done in double.  But a 1e-7/cell change in the
position would be a few*1e-8 change relative to the width of the
annulus.  If we were unlucky and that happened right around 0,
then a particle could be missed as the LC swept past.

Changing to 1e-6 should fix this, while still keeping the tolerances
tight, so that particles that have their LC bits cleared on the periodic
wrap don't get added from way behind the LC.  We may get a few of these;
this choice allows for a cell/1e5 strip that when it falls at the box
edge would double count particles that jump across right when the LC
crosses.  Very rare!