Increase MAL gap size to match page, and increase PAGE_SIZE to match hugepages.
The lattice IC step benefits enormously from a larger MAL gap size, which used
to be 512 particles. From a NUMA perspective, we would like gaps to reside on
the node of their thread, and the easiest way to do that is to make each gap
page-sized. But what page size? We've been using 4096, but actually most of
our machines probably use 2M hugepages. That's a big jump, but early tests
show no problems. Let's make 2M the default, but we'll want to watch this
closely.