light cones: Add --disable-lc-aux-bits and USE_LC_AUX_BITS
We use aux bits to track when particles cross a LC to avoid double
counting when a particle lands on the LC surface. But if we're
using many LC observers, e.g. tiling a box to form an all-sky LC,
then we might want to accept the rare duplicates rather than try
to extend the aux to potentially hundreds of bits.