light cone: add CheckLCAcrossWrap parameter to check for LC intersections with particles coming from across the periodic wrap
In the existing LC implementation, we accept that the edges of the
box will be "dirty", because a particle that is a cell across the
box may have an intersection with the LC by crossing the periodic
wrap. For simplicity, we weren't checking cells across the wrap.

In this PR, we add the option to consider a boundary layer that is one
cell deep in +/- x, y, and z. Effectively, when we do the for-loop
over cells, we change the loop from [0,cpd) to [-1,cpd+1). Then we
only accept intersections whose interpolated positions are within
the primary periodic image. This seems to work very well and fixes
the noticeable roll-off in LC particle density near the box edges.

This scheme can be used with or without LC aux bits. However, when
aux bits are in use, we ignore them when considering the boundary
layer. This is because we reset LC aux bits when a particle crosses
the wrap, but that won't happen until the merge. In rare instances
where a particle's interpolated position crosses the wrap and causes
it to land on the light cone surface, it may end up in the LC twice.
One can use CheckLCAcrossWrap = 0 to get the old behavior.