pre-release: add license to repo
Licensing Abacus under GPL-3.0-or-later.

A license notice header has been added to core source files as well, which isn't
strictly necessary but is considered best practice. I didn't add the notice to
docs, build system, or parameter files. They remain covered by the license, though.

There are a few non-GPL sources in the repo, but to the best of my knowledge, they
are all GPL-compatible (and the license notices have been retained).